+++ apps/admin/app/api/play-to-win/[id]/response.ts
@@ -1,0 +1,16 @@
+import { Prisma } from "@prisma/client";
+
+/**
+ * @swagger
+ * components:
+ *   schemas:
+ *     ConventionResponse:
+ *       type: object
+ *       properties:
+ *         convention:
+ *          type: object
+ *          $ref: '#/components/schemas/Convention'
+ */
+export interface ConventionResponse {
+  convention: Prisma.ConventionGetPayload<{}>
+}
