+++ apps/admin/app/api/play-to-win/edit/[id]/route.ts
@@ -2,6 +2,60 @@ import { IPlayToWinRequest } from "@/app/api/requests/play-to-win-request";
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/app/lib/prisma";
 
+/**
+ * @swagger
+ * /api/play-to-win/edit/{id}:
+ *   put:
+ *     tags:
+ *       - Play to Win Games
+ *     summary: Updates an existing play to win game
+ *     description: Updates the details of an existing play to win game based on the provided ID
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: integer
+ *         description: The unique identifier of the play to win game
+ *     requestBody:
+ *       description: Play to win game data to be updated
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             $ref: '#/components/schemas/PlayToWinRequest'
+ *     responses:
+ *       200:
+ *         description: Successfully updated the play to win game
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 message:
+ *                   type: string
+ *                   description: Success message after successfully updating the play to win game
+ *       400:
+ *         description: Invalid Data / Bad Request
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 error:
+ *                   type: string
+ *                   description: Error message indicating why the request failed
+ *       404:
+ *         description: Play to win game not found
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 error:
+ *                   type: string
+ *                   description: Error message indicating the play to win game was not found
+ */
 export async function PUT(request: NextRequest, { params }: { params: { id: string } }) {
   const playToWinItem: IPlayToWinRequest = await request.json();
